Tweak structure
