+++ packages/bot/src/common/services/logger.service/open-telemetry/processors/initializers/index.ts
@@ -1,0 +1,1 @@
+export * from "./rollem";
